Run CI with default cargo features (incl. `nvidia`), drop `ci`, `benchmark` features (#2782)
* Remove the `benchmarks` feature flag

We've removed the last usage of it back in 2020 in a9514e0268ec805c6d47.

Now _default features_ == _all features_, so there's no longer need to pass `--all-features` if you want to compile everything. Only exception is the `offline` feature of `visual-config`, but that actually removes functionality rather than adding it.

* Run CI with default cargo features, drop `ci` feature

The only addition in `default` on top of `ci` was the `nvidia` feature, and it turns out CI passes just fine with it enabled.

Also clean up the `[features]` sections of various `Cargo.toml`s.